feat: edit domain column name
