--- src/main/java/com/neoga/autionnara/bid/domain/Bid.java
@@ -1,25 +1,0 @@
-package com.neoga.autionnara.bid.domain;
-
-import com.neoga.autionnara.item.domain.Item;
-import com.neoga.autionnara.memberstore.member.domain.Members;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Entity
-public class Bid {
-    @Column(name="bid_id")
-    @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Members members;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id")
-    private Item item;
-    private int price;
-}
