--- src/main/java/com/neoga/autionnara/category/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.neoga.autionnara.category.repository;
-
-import com.neoga.autionnara.category.domain.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
