--- src/main/java/com/neoga/autionnara/category/service/CategoryService.java
@@ -1,9 +1,0 @@
-package com.neoga.autionnara.category.service;
-
-import com.neoga.autionnara.category.domain.Category;
-
-import java.util.List;
-
-public interface CategoryService {
-    List<Category> getCategoryList();
-}
