--- src/main/java/com/neoga/autionnara/category/service/CategoryServiceImpl.java
@@ -1,20 +1,0 @@
-package com.neoga.autionnara.category.service;
-
-import com.neoga.autionnara.category.domain.Category;
-import com.neoga.autionnara.category.repository.CategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class CategoryServiceImpl implements CategoryService {
-
-    @Autowired
-    private CategoryRepository categoryRepository;
-
-    @Override
-    public List<Category> getCategoryList() {
-        return categoryRepository.findAll();
-    }
-}
