--- src/main/java/com/neoga/autionnara/chat/domain/Message.java
@@ -1,29 +1,0 @@
-package com.neoga.autionnara.chat.domain;
-
-import com.neoga.autionnara.memberstore.member.domain.Members;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Entity
-public class Message {
-    @Column(name="message_id")
-    @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
-    private String content;
-    @Column(name="create_date")
-    private LocalDateTime createTime;
-    @Column(name="is_read")
-    private boolean isRead;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_id")
-    private Room room;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="member_id")
-    private Members sender;
-}
