--- src/main/java/com/neoga/autionnara/config/CommonBean.java
@@ -1,14 +1,0 @@
-package com.neoga.autionnara.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class CommonBean {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-}
