--- src/main/java/com/neoga/autionnara/image/repository/ItemImageRepository.java
@@ -1,4 +1,0 @@
-package com.neoga.autionnara.image.repository;
-
-public class ItemImageRepository {
-}
