--- src/main/java/com/neoga/autionnara/image/repository/MemberImageRepository.java
@@ -1,4 +1,0 @@
-package com.neoga.autionnara.image.repository;
-
-public class MemberImageRepository {
-}
