--- src/main/java/com/neoga/autionnara/item/repository/ItemRepository.java
@@ -1,7 +1,0 @@
-package com.neoga.autionnara.item.repository;
-
-import com.neoga.autionnara.item.domain.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
-}
