+++ src/main/java/com/neoga/boltauction/bid/domain/Bid.java
@@ -1,0 +1,25 @@
+package com.neoga.boltauction.bid.domain;
+
+import com.neoga.boltauction.item.domain.Item;
+import com.neoga.boltauction.memberstore.member.domain.Members;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Entity
+public class Bid {
+    @Column(name="id")
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long id;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Members members;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "item_id")
+    private Item item;
+    private int price;
+}
