+++ src/main/java/com/neoga/boltauction/category/controller/CategoryController.java
@@ -1,0 +1,23 @@
+package com.neoga.boltauction.category.controller;
+
+import com.neoga.boltauction.category.domain.Category;
+import com.neoga.boltauction.category.service.CategoryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/category")
+public class CategoryController {
+
+    private final CategoryService categoryService;
+
+    @GetMapping
+    public List<Category> getCategory() {
+        return categoryService.getCategoryList();
+    }
+}
