+++ src/main/java/com/neoga/boltauction/category/domain/Category.java
@@ -1,0 +1,17 @@
+package com.neoga.boltauction.category.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Entity
+public class Category {
+    @Column(name="id")
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private Long id;
+    private String name;
+}
