+++ src/main/java/com/neoga/boltauction/category/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package com.neoga.boltauction.category.repository;
+
+import com.neoga.boltauction.category.domain.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
