+++ src/main/java/com/neoga/boltauction/category/service/CategoryService.java
@@ -1,0 +1,9 @@
+package com.neoga.boltauction.category.service;
+
+import com.neoga.boltauction.category.domain.Category;
+
+import java.util.List;
+
+public interface CategoryService {
+    List<Category> getCategoryList();
+}
