+++ src/main/java/com/neoga/boltauction/category/service/CategoryServiceImpl.java
@@ -1,0 +1,20 @@
+package com.neoga.boltauction.category.service;
+
+import com.neoga.boltauction.category.domain.Category;
+import com.neoga.boltauction.category.repository.CategoryRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CategoryServiceImpl implements CategoryService {
+
+    @Autowired
+    private CategoryRepository categoryRepository;
+
+    @Override
+    public List<Category> getCategoryList() {
+        return categoryRepository.findAll();
+    }
+}
