+++ src/main/java/com/neoga/boltauction/chat/domain/Room.java
@@ -1,0 +1,18 @@
+package com.neoga.boltauction.chat.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+public class Room {
+    @Column(name="id")
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private Long id;
+    private LocalDateTime createDate;
+}
