+++ src/main/java/com/neoga/boltauction/item/repository/ItemRepository.java
@@ -1,0 +1,7 @@
+package com.neoga.boltauction.item.repository;
+
+import com.neoga.boltauction.item.domain.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
