+++ src/main/java/com/neoga/boltauction/memberstore/member/domain/Role.java
@@ -1,0 +1,5 @@
+package com.neoga.boltauction.memberstore.member.domain;
+
+public enum Role {
+    ROLE_MEMBER, ROLE_ADMIN
+}
