+++ src/test/java/com/neoga/boltauction/AutionnaraApplicationTests.java
@@ -1,0 +1,13 @@
+package com.neoga.boltauction;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AutionnaraApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
